+++ test/unit/couch-db-integration.test.js
@@ -1,6 +1,8 @@
 /**
  * this test checks the integration with couchdb
  * run 'npm run test:couchdb' to use it
+ * You need a running couchdb-instance on port 5984
+ * Run 'npm run couch:start' to spawn a docker-container
  */
 import assert from 'assert';
 import AsyncTestUtil from 'async-test-util';